Updated the Algae idle command to be running as the defualt in RobotContainer. Changed the XBoxController Class to CommandXboxController which allows for commands to be run automatically without having to call the specific button map on the within the xbox controller map and use JoystickButton to accomplish that. In addition to the neatness this permitted the triggers on the xbox controllers to have a native .whileTrue function attached to the instance of the controller.
Set up the basic autonomous routine to possibly spit out some coral after the robot stops moving.

Did a TON of refactoring of the code for neatness and readability. Added lots of doc strings to get things all sorted out. Hopefully this doesn't break anything major. If it does, see the first bit to revert to the main reaon to have these changes.